Move DynTrans to components/edge_convolutions.py
The EdgeConv family (DynEdgeConv, EdgeConvTito, DynTrans) is now
complete in its own module.
